requirements.txt
pynput==1.6.8
six==1.14.0
websocket-client==0.57.0
